Create docs for each of the new purpose groups
This is a first attempt at creating the content that could be on the new purpose pages (Observables, Views, etc.).

For the pages that I could, I copied the content from the Technical Highlights page and used that; for the others, I listed the modules by their collection, similar to what’s on the API Docs page.

I think we’ll want to spend more time fleshing out what should be on these pages, but at least this is a start.

Related to https://github.com/canjs/canjs/issues/3645