Add can-namespace docs to this repo
The `can-namespace` docs needed to be updated with their new @parent and @collection, but we can’t make a new release of `can-namespace`, so I’m ignoring the npm module and adding the docs into the repo.

The GitHub stars button shows the stars for `can` instead of `can-namespace`, but everything else seems ok.